Moved files in domain folder
