+++ src/components/Gesture.svelte
@@ -21,8 +21,8 @@
   import ImageSkeleton from './skeletonloading/ImageSkeleton.svelte';
   import GestureTilePart from './GestureTilePart.svelte';
   import StaticConfiguration from '../StaticConfiguration';
-  import Gesture from '../script/domain/Gesture';
-  import { classifier, gestures, liveAccelerometerData } from '../script/stores/Stores';
+  import { gestures, liveAccelerometerData } from '../script/stores/Stores';
+  import Gesture from '../script/domain/stores/gesture/Gesture';
 
   // Variables for component
   export let onNoMicrobitSelect: () => void;
