+++ src/pages/filter/D3Plot.svelte
@@ -9,11 +9,11 @@
   import { get } from 'svelte/store';
   import * as d3 from 'd3';
   import { state } from '../../script/stores/uiStore';
-  import { getPrevData } from '../../script/stores/mlStore';
+  import { RecordingData, getPrevData } from '../../script/stores/mlStore';
   import { gestures, liveAccelerometerData } from '../../script/stores/Stores';
   import FilterTypes, { FilterType } from '../../script/domain/FilterTypes';
   import FilterGraphLimits from '../../script/utils/FilterLimits';
-  import { GestureData } from '../../script/domain/Gesture';
+  import { GestureData } from '../../script/domain/stores/gesture/Gesture';
 
   export let filterType: FilterType;
   export let fullScreen: boolean = false;
@@ -165,7 +165,7 @@
       if (!classes.includes(gestureClass)) {
         classes.push(gestureClass);
       }
-      gestureClassObject.recordings.map(recording => {
+      gestureClassObject.recordings.map((recording: RecordingData) => {
         const ID = recording.ID;
         const x = filterFunction(recording.data.x);
         const y = filterFunction(recording.data.y);
