+++ src/script/domain/stores/gesture/Gesture.ts
@@ -4,12 +4,12 @@
  * SPDX-License-Identifier: MIT
  */
 import { Readable, Subscriber, Unsubscriber, Writable, derived, get } from 'svelte/store';
-import { RecordingData } from '../stores/mlStore';
-import { PersistantGestureData } from './Gestures';
-import { PinTurnOnState } from '../../components/output/PinSelectorUtil';
-import MBSpecs from '../microbit-interfacing/MBSpecs';
-import BindableValue from './BindableValue';
 import GestureConfidence from './GestureConfidence';
+import { PersistantGestureData } from './Gestures';
+import MBSpecs from '../../../microbit-interfacing/MBSpecs';
+import { PinTurnOnState } from '../../../../components/output/PinSelectorUtil';
+import { RecordingData } from '../../../stores/mlStore';
+import BindableValue from '../BindableValue';
 
 export type GestureID = number;
