+++ src/script/domain/stores/gesture/GestureConfidence.ts
